+++ kjr_grant/controllers/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import portal
 from . import website
+from . import legacy_redirects